# apps/web/openapi/petclinic-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Petclinic MSS301 — Aggregated API",
    "version": "1.0.0",
    "description": "Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews"
  },
  "servers": [
    {
      "url": "http://localhost:8180",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/vets/{vetId}/ratings": {
      "get": {
        "tags": ["Vet Ratings"],
        "summary": "List ratings of a vet (paginated)",
        "description": "Vet không tồn tại → 404. Use ?page=0&size=20&sort=rateDate,desc. Optional ?year=YYYY lọc rating theo năm của rateDate.",
        "operationId": "listVetRatings",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageRatingResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Vet Ratings"],
        "summary": "Add or update rating of a vet (UPSERT)",
        "description": "score 1-5 (out of range → 400 ProblemDetail). customerName lấy từ JWT claim 'username' (auth-service ký vào access token) — client không thể giả mạo identity. Vet không tồn tại → 404. JWT thiếu claim 'username' → 400 (token version cũ → cần login lại). POST lần thứ 2 của cùng customer trên cùng vet → UPSERT (update rating cũ + đổi rateDate). Unique (vet_id, customer_name) ở DB layer. Body field customerName (nếu có) bị Jackson silently ignore.",
        "operationId": "addVetRating",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vets/{vetId}/ratings/summary": {
      "get": {
        "tags": ["Vet Ratings"],
        "summary": "Get rating summary of a vet",
        "description": "Trả {count, average (null nếu count=0), distribution Map<1..5, count>}. Distribution luôn có đủ 5 key, value=0 nếu không có rating ở score đó.",
        "operationId": "getVetRatingsSummary",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingSummaryResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vets/{vetId}/ratings/{ratingId}": {
      "delete": {
        "tags": ["Vet Ratings"],
        "summary": "Delete rating (rating tồn tại nhưng thuộc vet khác → 404, không leak)",
        "operationId": "deleteVetRating",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "ratingId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/reviews/{id}": {
      "get": {
        "tags": ["Reviews"],
        "summary": "Get review detail",
        "operationId": "getReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Reviews"],
        "summary": "USER soft-delete review của mình — status → DELETED",
        "operationId": "deleteReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      },
      "patch": {
        "tags": ["Reviews"],
        "summary": "USER chỉnh sửa review của mình — trong 7 ngày kể từ tạo",
        "operationId": "updateReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{id}/unhide": {
      "patch": {
        "tags": ["Reviews Moderation"],
        "summary": "Khôi phục HIDDEN → PUBLISHED (ADMIN only)",
        "operationId": "unhideReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{id}/hide": {
      "patch": {
        "tags": ["Reviews Moderation"],
        "summary": "Ẩn review (kèm reason audit) — status → HIDDEN",
        "operationId": "hideReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HideReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{id}/approve": {
      "patch": {
        "tags": ["Reviews Moderation"],
        "summary": "Duyệt PENDING_MODERATION hoặc HIDDEN → PUBLISHED",
        "operationId": "approveReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/reviews": {
      "get": {
        "tags": ["Reviews Moderation"],
        "summary": "List queue moderation — default status=PENDING_MODERATION",
        "operationId": "listPendingReviews",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "PENDING_MODERATION",
              "enum": ["PUBLISHED", "PENDING_MODERATION", "FLAGGED", "HIDDEN", "DELETED"]
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageReviewResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/reviews/{id}": {
      "delete": {
        "tags": ["Reviews Moderation"],
        "summary": "Hard moderate — soft-delete row (ADMIN only)",
        "operationId": "adminDeleteReview",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "unpaged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
...
      "RatingRequest": {
        "type": "object",
        "properties": {
          "score": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "predefinedDescription": {
            "type": "string",
            "enum": ["POOR", "AVERAGE", "GOOD", "EXCELLENT"]
          }
        },
        "required": ["score"]
      },
      "RatingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "vetId": {
            "type": "integer",
            "format": "int64"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "predefinedDescription": {
            "type": "string",
            "enum": ["POOR", "AVERAGE", "GOOD", "EXCELLENT"]
          },
          "customerName": {
            "type": "string"
          },
          "rateDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PageRatingResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RatingResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "RatingSummaryResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "average": {
            "type": "number",
            "format": "double"
          },
          "distribution": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int64"
            }
          }
        }
      },
...
      "ReviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "targetType": {
            "type": "string",
            "enum": ["VET", "PRODUCT", "VISIT"]
          },
          "targetId": {
            "type": "integer",
            "format": "int64"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "authorName": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["PUBLISHED", "PENDING_MODERATION", "FLAGGED", "HIDDEN", "DELETED"]
          },
          "helpfulCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifiedDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateReviewRequest": {
        "type": "object",
        "properties": {
          "rating": {
            "type": "integer",
            "format": "int32",
            "maximum": 5,
            "minimum": 1
          },
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "comment": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 0
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "HideReviewRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 0
          }
        },
        "required": ["reason"]
      },
      "PageReviewResponse": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {}
  }
}